Правки по комментам
