+++ test_generator/library/interface_content.py
@@ -9,9 +9,9 @@ class InterfaceContent(str, Enum):
         url = self._api_url + f"$path"
         params = {
             **self._actor.params,
-            CHANGE_ME_TO_NEED_PARAMS
+            # CHANGE_ME_TO_NEED_PARAMS
         }
-
+        raise NotImplementedError('Need change parameters/body in new generated interface')
         return await API.get(url, params=params, headers=self._actor.metadata)
         """
     POST = """
@@ -19,13 +19,13 @@ class InterfaceContent(str, Enum):
         url = self._api_url + f"$path"
         params = {
             **self._actor.params,
-            CHANGE_ME_TO_NEED_PARAMS
+            # CHANGE_ME_TO_NEED_PARAMS
         }
         
         data = {
-            CHANGE_ME_TO_NEED_BODY
+            # CHANGE_ME_TO_NEED_BODY
         }
-
+        raise NotImplementedError('Need change parameters/body in new generated interface')
         return await API.post(url, params=params, data=data, headers=self._actor.metadata)
         """
     DELETE = """
@@ -33,26 +33,23 @@ class InterfaceContent(str, Enum):
         url = self._api_url + f"$path"
         params = {
             **self._actor.params,
-            CHANGE_ME_TO_NEED_PARAMS
+            # CHANGE_ME_TO_NEED_PARAMS
         }
-
+        raise NotImplementedError('Need change parameters/body in new generated interface')
         return await API.delete(url, params=params, headers=self._actor.metadata)
         """
 
     @classmethod
-    def fill_get_template(cls, schema_data: SchemaData) -> str:
-        return (InterfaceContent.GET.replace('$path', schema_data.path)
-                                    .replace('$name', schema_data.interface_method)
-                                    .replace('$params', ', '.join(schema_data.args)))
-
-    @classmethod
-    def fill_post_template(cls, schema_data: SchemaData) -> str:
-        return (InterfaceContent.POST.replace('$path', schema_data.path)
-                                     .replace('$name', schema_data.interface_method)
-                                     .replace('$params', ', '.join(schema_data.args)))
-
-    @classmethod
-    def fill_delete_template(cls, schema_data: SchemaData) -> str:
-        return (InterfaceContent.DELETE.replace('$path', schema_data.path)
-                                       .replace('$name', schema_data.interface_method)
-                                       .replace('$params', ', '.join(schema_data.args)))
+    def fill_template(cls, schema_data: SchemaData) -> str:
+        match schema_data.http_method.upper():
+            case 'GET':
+                func = InterfaceContent.GET
+            case 'POST':
+                func = InterfaceContent.POST
+            case 'DELETE':
+                func = InterfaceContent.DELETE
+            case default:
+                raise RuntimeError(f'{schema_data.http_method} is not supported')
+        return (func.replace('$path', schema_data.path)
+                    .replace('$name', schema_data.interface_method)
+                    .replace('$params', ', '.join(schema_data.args)))
